# swagger_api_docs.json
{"openapi":"3.0.3","info":{"title":"solana_todo_backend","description":"REST API backend for Solana Todo application","contact":{"name":"Your Name","email":"your.email@example.com"},"license":{"name":""},"version":"0.1.0"},"paths":{"/api/auth":{"post":{"tags":["auth"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"Successfully authenticated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"401":{"description":"Authentication failed"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}}}},"/api/todos":{"get":{"tags":["todo"],"operationId":"get_todos","responses":{"200":{"description":"List of todos","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoResponse"}}}}},"401":{"description":"Unauthorized"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}},"security":[{"bearer_auth":[]}]},"post":{"tags":["todo"],"operationId":"create_todo","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoRequest"}}},"required":true},"responses":{"201":{"description":"Todo created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}},"security":[{"bearer_auth":[]}]}},"/api/todos/{id}":{"put":{"tags":["todo"],"operationId":"update_todo","parameters":[{"name":"id","in":"path","description":"Todo ID","required":true,"schema":{"type":"integer","format":"int64","minimum":0}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoRequest"}}},"required":true},"responses":{"200":{"description":"Todo updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoResponse"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Todo not found"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}},"security":[{"bearer_auth":[]}]},"delete":{"tags":["todo"],"operationId":"delete_todo","parameters":[{"name":"id","in":"path","description":"Todo ID","required":true,"schema":{"type":"integer","format":"int64","minimum":0}}],"responses":{"204":{"description":"Todo deleted successfully"},"401":{"description":"Unauthorized"},"404":{"description":"Todo not found"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}},"security":[{"bearer_auth":[]}]}},"/api/transactions/prepare/create":{"post":{"tags":["transaction"],"operationId":"prepare_create_transaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoRequest"}}},"required":true},"responses":{"200":{"description":"Transaction prepared successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreparedTransaction"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}},"security":[{"bearer_auth":[]}]}},"/api/transactions/prepare/delete":{"post":{"tags":["transaction"],"operationId":"prepare_delete_transaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteTodoTransactionRequest"}}},"required":true},"responses":{"200":{"description":"Transaction prepared successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreparedTransaction"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}},"security":[{"bearer_auth":[]}]}},"/api/transactions/prepare/update/{id}":{"post":{"tags":["transaction"],"operationId":"prepare_update_transaction","parameters":[{"name":"id","in":"path","description":"Todo ID","required":true,"schema":{"type":"integer","format":"int64","minimum":0}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoRequest"}}},"required":true},"responses":{"200":{"description":"Transaction prepared successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreparedTransaction"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}},"security":[{"bearer_auth":[]}]}},"/api/transactions/submit":{"post":{"tags":["transaction"],"operationId":"submit_transaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignedTransaction"}}},"required":true},"responses":{"200":{"description":"Transaction submitted successfully","content":{"text/plain":{"schema":{"type":"string"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"429":{"description":"Rate limit exceeded"},"500":{"description":"Internal server error"}},"security":[{"bearer_auth":[]}]}}},"components":{"schemas":{"AuthRequest":{"type":"object","required":["public_key","signature","timestamp"],"properties":{"public_key":{"type":"string","example":"HXtBm8XZbxaTt41uqaKhwUAa6Z1aPyvJdsZVENiWsetg"},"signature":{"type":"string","example":"3AuheKDvzxG6QM4gQHPNFTQ5wGz3aEZnECJK5Lp1e5orTAsyUNZKrGZq25e3XPQZiLgVj7LNzjwERFvxdL4Zx54M"},"timestamp":{"type":"integer","format":"int64","example":"1625097600"}}},"AuthResponse":{"type":"object","required":["token","expires_in","public_key"],"properties":{"expires_in":{"type":"integer","format":"int64","example":"86400","minimum":0},"public_key":{"type":"string","example":"HXtBm8XZbxaTt41uqaKhwUAa6Z1aPyvJdsZVENiWsetg"},"token":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}},"CreateTodoRequest":{"type":"object","required":["description","due_date"],"properties":{"description":{"type":"string","example":"Complete the Solana project","maxLength":280,"minLength":1},"due_date":{"type":"integer","format":"int64","example":"1625097600"}}},"DateTime<Utc>":{"type":"string","format":"date-time","description":"ISO 8601 date and time","example":"2024-03-03T18:35:30Z"},"DeleteTodoTransactionRequest":{"type":"object","required":["todo_id"],"properties":{"todo_id":{"type":"integer","format":"int64","example":"1","minimum":0}}},"PreparedTransaction":{"type":"object","required":["serialized_transaction","transaction_type"],"properties":{"metadata":{"type":"string","example":"{\"todo_id\":1,\"description\":\"Buy groceries\"}","nullable":true},"serialized_transaction":{"type":"string","example":"AQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAAEDyjtJ6Jf8zd9P/YoVvMM9Yg9n/vYMzjBFtKyI5IVwBGNBxiqYhI0UXeG9LrYPk4p8lZCLXKJTa6nQ61KJ1gAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA4O5d4OTMrMvbCpVv5maDe82SnNxl6OYOHAMBIwXxF0PS7UOi9/tQYZcG9zV9k5OTcUXJIcI+iRGhcGQly3yZAgAAAAAHYUJnTxU6m3Fck70YwVzLpNcbJ0yKVpzQzLcR7y9ZVQEDAwABAEoAAAAAAAAAjbv2dF80rSTn3xTrH9Y1XRdgFxvLhd+jN2gDBQD/GfwMjinlkDl90P7xO3YKBp9BQkM/fqGcJ5aoVaA8tzoBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"},"transaction_type":{"type":"string","example":"create_todo"}}},"SignedTransaction":{"type":"object","required":["signature","serialized_transaction"],"properties":{"serialized_transaction":{"type":"string","example":"AQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABAAEDyjtJ6Jf8zd9P/YoVvMM9YoFnVvYMzjBFtKyI5IVwBGNBxiqYhI0UXeG9LrYPk4p8lZCLXKJTa6nQ61KJ1gAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA4O5d4OTMrMvbCpVv5maDe82SnNxl6OYOHAMBIwXxF0PS7UOi9/tQYZcG9zV9k5OTcUXJIcI+iRGhcGQly3yZAgAAAAAHYUJnTxU6m3Fck70YwVzLpNcbJ0yKVpzQzLcR7y9ZVQEDAwABAEoAAAAAAAAAjbv2dF80rSTn3xTrH9Y1XRdgFxvLhd+jN2gDBQD/GfwMjinlkDl90P7xO3YKBp9BQkM/fqGcJ5aoVaA8tzoBAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"},"signature":{"type":"string","example":"3WZ4sBJm5eGzRfRXLCuUktaE7xMW25dveKfpGxCCkyvMfPQiJKgf1nuFCvhKTQUCHYKGEE6KuQnUijBTKMhNcWd7"}}},"TodoResponse":{"type":"object","required":["id","description","completed","due_date","owner","created_at"],"properties":{"completed":{"type":"boolean","example":"false"},"created_at":{"$ref":"#/components/schemas/DateTimeWrapper"},"description":{"type":"string","example":"Complete the Solana project"},"due_date":{"type":"integer","format":"int64","example":"1625097600"},"id":{"type":"integer","format":"int64","example":"1","minimum":0},"owner":{"type":"string","example":"HXtBm8XZbxaTt41uqaKhwUAa6Z1aPyvJdsZVENiWsetg"},"updated_at":{"allOf":[{"$ref":"#/components/schemas/DateTimeWrapper"}],"nullable":true}}},"UpdateTodoRequest":{"type":"object","properties":{"completed":{"type":"boolean","example":"true","nullable":true},"description":{"type":"string","example":"Complete the Solana project updated","nullable":true,"maxLength":280,"minLength":1}}}},"securitySchemes":{"bearer_auth":{"type":"apiKey","in":"header","name":"Authorization"}}},"tags":[{"name":"auth","description":"Authentication endpoints"},{"name":"todos","description":"Todo management endpoints"},{"name":"transactions","description":"Transaction preparation and submission endpoints"}]}